S.L Przeniesiona na VS2k5 F-cje deprecated na razie rzucaja warningami. Do poprawienia w przyszłości (najlepiej przez zestaw f-cji inline dla != vs2k5) Poprawione odwołania do ścieżek. W depends siedzą teraz wszystkie potrzebne biblioteki (wg. ścieżek w projektach)
git-svn-id: http://www.konnekt.info/svn/staminalib/trunk@223 e7532c55-18f9-0310-997f-f942ee4ad645

+++ Stamina/Time64.cpp
@@ -50,6 +50,7 @@ namespace Stamina {
 		if (tmp) {
 			timer = *tmp;
 		}
+
 		return timer;
 	}
 
@@ -110,9 +111,9 @@ namespace Stamina {
 
 
 	std::string Time64::getTimeString(const char* hourStr, const char* minStr, const char* secStr, bool needHour, bool needMin, bool needSec) const {
-		int hour = this->sec / (60*60);
-		int min = (this->sec % (60*60)) / 60;
-		int sec = (this->sec % (60));
+		int hour = (int)this->sec / (60*60);
+		int min = (int)(this->sec % (60*60)) / 60;
+		int sec = (int)(this->sec % (60));
 		std::string str;
 		char buff [20];
 		if (hour || needHour) {
